requirements/requirements.txt
openai
python-dotenv
requests
typer
